moves agreement events to a subpackage agreement
