+++ webofneeds/won-bot/src/main/java/won/bot/framework/eventbot/event/impl/analyzation/agreement/AgreementAcceptedEvent.java
@@ -1,4 +1,4 @@
-package won.bot.framework.eventbot.event.impl.analyzation;
+package won.bot.framework.eventbot.event.impl.analyzation.agreement;
 
 import won.protocol.model.Connection;
