+++ webofneeds/won-bot/src/main/java/won/bot/framework/eventbot/event/impl/analyzation/agreement/AgreementEvent.java
@@ -1,4 +1,4 @@
-package won.bot.framework.eventbot.event.impl.analyzation;
+package won.bot.framework.eventbot.event.impl.analyzation.agreement;
 
 import won.bot.framework.eventbot.event.BaseNeedAndConnectionSpecificEvent;
 import won.protocol.model.Connection;
